Move service module functions into dedicated modules
This commit reorganizes the tansu-service codebase by moving
functionality from the `stream.rs` module into dedicated modules. The
changes include:

- Extract frame/channel/api functionality into separate modules - Clean
up imports and remove unused code - Centralize error handling and
metrics - Move TCP-related code to stream.rs

The goal is to improve code organization and make the codebase more
maintainable.
