+++ tansu-service/src/api.rs
@@ -12,257 +12,184 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{
-    collections::BTreeMap,
-    error::Error,
-    fmt::{self, Debug, Display},
-    sync::{Arc, LazyLock, Mutex},
-    time::SystemTime,
-};
+use std::{collections::BTreeMap, marker::PhantomData, sync::Arc};
 
-use bytes::Bytes;
-use opentelemetry::{KeyValue, metrics::Histogram};
-use rama::{Context, Layer, Service, context::Extensions, error::BoxError, matcher::Matcher};
-use tansu_sans_io::{Body, Frame, Header};
+use rama::{Context, Service, service::BoxService};
+use tansu_sans_io::{
+    ApiKey, ApiVersionsRequest, ApiVersionsResponse, Body, ErrorCode, Frame, Header,
+    RootMessageMeta, api_versions_response::ApiVersion,
+};
 use tracing::debug;
 
-use crate::{METER, Result};
+use crate::Error;
 
-pub mod api_version;
-pub mod describe_config;
-pub mod metadata;
-pub mod produce;
-
-#[derive(Copy, Clone, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct ApiKey(pub i16);
-
-impl AsRef<i16> for ApiKey {
-    fn as_ref(&self) -> &i16 {
-        &self.0
-    }
+// An versions service with a supported set of APIs
+#[derive(Clone, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
+pub struct ApiVersionsService<E> {
+    supported: Vec<i16>,
+    error: PhantomData<E>,
 }
 
-pub type ApiVersion = i16;
-pub type CorrelationId = i32;
-pub type ClientId = Option<String>;
-
-#[derive(Clone, Debug, PartialEq, PartialOrd)]
-pub struct ApiRequest {
-    pub api_key: ApiKey,
-    pub api_version: ApiVersion,
-    pub correlation_id: CorrelationId,
-    pub client_id: ClientId,
-    pub body: Body,
-}
-
-impl TryFrom<ApiRequest> for Bytes {
-    type Error = BoxError;
-
-    fn try_from(api_request: ApiRequest) -> Result<Self, Self::Error> {
-        Frame::request(
-            Header::Request {
-                api_key: api_request.api_key.0,
-                api_version: api_request.api_version,
-                correlation_id: api_request.correlation_id,
-                client_id: api_request.client_id,
-            },
-            api_request.body,
-        )
-        .map_err(Into::into)
-    }
-}
-
-impl<State> Matcher<State, ApiRequest> for ApiKey
+impl<State, E> Service<State, ApiVersionsRequest> for ApiVersionsService<E>
 where
-    State: Debug,
+    State: Clone + Send + Sync + 'static,
+    E: std::error::Error + Send + Sync + 'static,
 {
-    fn matches(
-        &self,
-        ext: Option<&mut Extensions>,
-        ctx: &Context<State>,
-        req: &ApiRequest,
-    ) -> bool {
-        let _ = (ext, ctx);
-        debug!(api_key = self.0, ?req);
-        self.0 == req.api_key.0
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, PartialOrd)]
-pub struct ApiResponse {
-    pub api_key: ApiKey,
-    pub api_version: ApiVersion,
-    pub correlation_id: CorrelationId,
-    pub body: Body,
-}
-
-impl TryFrom<ApiResponse> for Bytes {
-    type Error = BoxError;
-
-    fn try_from(api_response: ApiResponse) -> Result<Self, Self::Error> {
-        Frame::response(
-            Header::Response {
-                correlation_id: api_response.correlation_id,
-            },
-            api_response.body,
-            api_response.api_key.0,
-            api_response.api_version,
-        )
-        .map_err(Into::into)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, PartialOrd)]
-struct FrameTypeError {
-    frame: Box<Frame>,
-}
-
-impl Display for FrameTypeError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "unexpected frame type")
-    }
-}
-impl Error for FrameTypeError {}
-
-pub fn read_api_request(buffer: Bytes) -> Result<ApiRequest> {
-    match Frame::request_from_bytes(&buffer[..])? {
-        Frame {
-            header:
-                Header::Request {
-                    api_key,
-                    api_version,
-                    correlation_id,
-                    client_id,
-                },
-            body,
-            ..
-        } => Ok(ApiRequest {
-            api_key: ApiKey(api_key),
-            api_version,
-            correlation_id,
-            client_id,
-            body,
-        }),
-
-        frame => Err(FrameTypeError {
-            frame: Box::new(frame),
-        }
-        .into()),
-    }
-}
-
-static API_RESPONSE_FROM_BYTES_DURATION: LazyLock<Histogram<u64>> = LazyLock::new(|| {
-    METER
-        .u64_histogram("api_response_from_bytes_duration")
-        .with_unit("ms")
-        .with_description("Latency of converting bytes into an API response in milliseconds")
-        .build()
-});
-
-pub fn read_api_response(
-    buffer: Bytes,
-    api_key: ApiKey,
-    api_version: ApiVersion,
-) -> Result<ApiResponse> {
-    let attributes = [KeyValue::new("api_key", api_key.0.to_string())];
-    let start = SystemTime::now();
-
-    match Frame::response_from_bytes(&buffer[..], api_key.0, api_version).inspect(|_| {
-        API_RESPONSE_FROM_BYTES_DURATION.record(
-            start
-                .elapsed()
-                .map_or(0, |duration| duration.as_millis() as u64),
-            &attributes,
-        )
-    })? {
-        Frame {
-            header: Header::Response { correlation_id },
-            body,
-            ..
-        } => Ok(ApiResponse {
-            api_key,
-            api_version,
-            correlation_id,
-            body,
-        }),
-
-        frame => Err(FrameTypeError {
-            frame: Box::new(frame),
-        }
-        .into()),
-    }
-}
-
-type KeyVersionService<S> = Arc<Mutex<BTreeMap<(ApiKey, ApiVersion), Arc<S>>>>;
-
-#[derive(Clone, Debug, Default)]
-pub struct ApiKeyVersionService<S> {
-    inner: KeyVersionService<S>,
-    model: S,
-}
-
-impl<S, State> Service<State, ApiRequest> for ApiKeyVersionService<S>
-where
-    S: Service<State, ApiRequest, Response = ApiResponse> + Clone,
-    S::Error: Into<BoxError> + Send + Debug + 'static,
-    State: Send + Sync + 'static,
-{
-    type Response = ApiResponse;
-    type Error = BoxError;
+    type Response = ApiVersionsResponse;
+    type Error = E;
 
     async fn serve(
         &self,
-        ctx: Context<State>,
-        req: ApiRequest,
+        _ctx: Context<State>,
+        _req: ApiVersionsRequest,
     ) -> Result<Self::Response, Self::Error> {
-        debug!(?req);
-
-        let service = self
-            .inner
-            .lock()
-            .map(|mut guard| {
-                let key = (req.api_key, req.api_version);
-
-                guard
-                    .entry(key)
-                    .or_insert(Arc::new(self.model.clone()))
-                    .to_owned()
-            })
-            .expect("poison");
-
-        service
-            .serve(ctx, req)
-            .await
-            .inspect(|response| debug!(?response))
-            .inspect_err(|err| debug!(?err))
-            .map_err(Into::into)
+        Ok::<_, E>(
+            ApiVersionsResponse::default()
+                .finalized_features(Some([].into()))
+                .finalized_features_epoch(Some(-1))
+                .supported_features(Some([].into()))
+                .zk_migration_ready(Some(false))
+                .error_code(ErrorCode::None.into())
+                .api_keys(Some(
+                    RootMessageMeta::messages()
+                        .requests()
+                        .iter()
+                        .filter(|(api_key, _)| self.supported.contains(api_key))
+                        .map(|(_, meta)| {
+                            ApiVersion::default()
+                                .api_key(meta.api_key)
+                                .min_version(meta.version.valid.start)
+                                .max_version(meta.version.valid.end)
+                        })
+                        .collect(),
+                ))
+                .throttle_time_ms(Some(0)),
+        )
     }
 }
 
-#[derive(Copy, Clone, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct ApiKeyVersionLayer;
+impl<State, E> Service<State, Body> for ApiVersionsService<E>
+where
+    State: Clone + Send + Sync + 'static,
+    E: std::error::Error + From<tansu_sans_io::Error> + Send + Sync + 'static,
+{
+    type Response = Body;
+    type Error = E;
 
-impl<S> Layer<S> for ApiKeyVersionLayer {
-    type Service = ApiKeyVersionService<S>;
+    async fn serve(&self, ctx: Context<State>, req: Body) -> Result<Self::Response, Self::Error> {
+        let req = ApiVersionsRequest::try_from(req)?;
+        self.serve(ctx, req).await.map(Into::into)
+    }
+}
 
-    fn layer(&self, model: S) -> Self::Service {
-        ApiKeyVersionService {
-            inner: Arc::new(Mutex::new(BTreeMap::new())),
-            model,
+impl<State, E> Service<State, Frame> for ApiVersionsService<E>
+where
+    State: Clone + Send + Sync + 'static,
+    E: std::error::Error + From<tansu_sans_io::Error> + Send + Sync + 'static,
+{
+    type Response = Frame;
+    type Error = E;
+
+    async fn serve(&self, ctx: Context<State>, req: Frame) -> Result<Self::Response, Self::Error> {
+        let correlation_id = req.correlation_id()?;
+        self.serve(ctx, req.body).await.map(|body| Frame {
+            size: 0,
+            header: Header::Response { correlation_id },
+            body,
+        })
+    }
+}
+
+/// Route frames to a service based via the API key
+#[derive(Clone, Debug, Default)]
+pub struct FrameRouteService<State = (), E = Error> {
+    routes: Arc<BTreeMap<i16, BoxService<State, Frame, Frame, E>>>,
+}
+
+impl<State, E> FrameRouteService<State, E>
+where
+    State: Clone + Send + Sync + 'static,
+    E: std::error::Error + From<tansu_sans_io::Error> + From<Error> + Send + Sync + 'static,
+{
+    pub fn new(routes: Arc<BTreeMap<i16, BoxService<State, Frame, Frame, E>>>) -> Self {
+        Self { routes }
+    }
+
+    pub fn builder() -> FrameRouteBuilder<State, E> {
+        FrameRouteBuilder::<State, E>::new()
+    }
+}
+
+impl<State, E> Service<State, Frame> for FrameRouteService<State, E>
+where
+    State: Clone + Send + Sync + 'static,
+    E: std::error::Error + From<tansu_sans_io::Error> + From<Error> + Send + Sync + 'static,
+{
+    type Response = Frame;
+    type Error = E;
+
+    async fn serve(&self, ctx: Context<State>, req: Frame) -> Result<Self::Response, Self::Error> {
+        debug!(?req);
+
+        let api_key = req.api_key()?;
+
+        if let Some(service) = self.routes.get(&api_key) {
+            service.serve(ctx, req).await
+        } else {
+            Err(E::from(Error::UnknownServiceFrame(Box::new(req))))
         }
     }
 }
 
-#[derive(Clone, Debug, PartialEq, PartialOrd)]
-pub enum UnexpectedApiBodyError {
-    Request(Box<ApiRequest>),
-    Response(Box<ApiResponse>),
+/// A frame route builder providing an API versions response with the available routes
+#[derive(Debug)]
+pub struct FrameRouteBuilder<State, E> {
+    routes: BTreeMap<i16, BoxService<State, Frame, Frame, E>>,
 }
 
-impl Display for UnexpectedApiBodyError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "unexpected response")
+impl<State, E> FrameRouteBuilder<State, E>
+where
+    State: Clone + Send + Sync + 'static,
+    E: std::error::Error + From<tansu_sans_io::Error> + Send + Sync + 'static,
+{
+    fn new() -> Self {
+        Self {
+            routes: BTreeMap::new(),
+        }
+    }
+
+    pub fn with_service<S>(self, service: S) -> Result<Self, Error>
+    where
+        S: Into<BoxService<State, Frame, Frame, E>> + ApiKey,
+    {
+        self.with_route(S::KEY, service.into())
+    }
+
+    pub fn with_route(
+        mut self,
+        api_key: i16,
+        service: BoxService<State, Frame, Frame, E>,
+    ) -> Result<Self, Error> {
+        self.routes
+            .insert(api_key, service)
+            .map_or(Ok(self), |_existing| Err(Error::DuplicateRoute(api_key)))
+    }
+
+    pub fn build(self) -> Result<FrameRouteService<State, E>, Error> {
+        let api_key = ApiVersionsRequest::KEY;
+        let mut supported = self.routes.keys().copied().collect::<Vec<_>>();
+        supported.push(api_key);
+
+        self.with_route(
+            api_key,
+            ApiVersionsService {
+                supported,
+                error: PhantomData,
+            }
+            .boxed(),
+        )
+        .map(|builder| FrameRouteService {
+            routes: Arc::new(builder.routes),
+        })
     }
 }
-
-impl Error for UnexpectedApiBodyError {}
