+++ tansu-service/tests/common/mod.rs
@@ -1,0 +1,81 @@
+// Copyright ⓒ 2024-2025 Peter Morgan <peter.james.morgan@gmail.com>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::{fmt, io, sync::Arc};
+
+use tracing::subscriber::DefaultGuard;
+use tracing_subscriber::{EnvFilter, filter::ParseError};
+
+#[derive(Clone, Debug, thiserror::Error)]
+pub(crate) enum Error {
+    #[allow(dead_code)]
+    Io(Arc<io::Error>),
+
+    #[allow(dead_code)]
+    Message(String),
+
+    #[allow(dead_code)]
+    ParseFilter(Arc<ParseError>),
+
+    Protocol(#[from] tansu_sans_io::Error),
+    Service(#[from] tansu_service::Error),
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{self:?}")
+    }
+}
+
+impl From<io::Error> for Error {
+    fn from(value: io::Error) -> Self {
+        Self::Io(Arc::new(value))
+    }
+}
+
+impl From<ParseError> for Error {
+    fn from(value: ParseError) -> Self {
+        Self::ParseFilter(Arc::new(value))
+    }
+}
+
+pub(crate) fn init_tracing() -> Result<DefaultGuard, Error> {
+    use std::{fs::File, sync::Arc, thread};
+
+    Ok(tracing::subscriber::set_default(
+        tracing_subscriber::fmt()
+            .with_level(true)
+            .with_line_number(true)
+            .with_thread_names(false)
+            .with_env_filter(
+                EnvFilter::from_default_env()
+                    .add_directive(format!("{}=debug", env!("CARGO_CRATE_NAME")).parse()?),
+            )
+            .with_writer(
+                thread::current()
+                    .name()
+                    .ok_or(Error::Message(String::from("unnamed thread")))
+                    .and_then(|name| {
+                        File::create(format!(
+                            "../logs/{}/{}::{name}.log",
+                            env!("CARGO_PKG_NAME"),
+                            env!("CARGO_CRATE_NAME")
+                        ))
+                        .map_err(Into::into)
+                    })
+                    .map(Arc::new)?,
+            )
+            .finish(),
+    ))
+}
